// SDKs/twilio.com/api/1.36.0/typescript/src/sdk/models/shared/apiv2010accountrecordingrecordingtranscription.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";
import { RecordingTranscriptionEnumStatusEnum } from "./recordingtranscriptionenumstatusenum";


export class ApiV2010AccountRecordingRecordingTranscription extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=api_version" })
  apiVersion?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: string;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: string;

  @Metadata({ data: "json, name=duration" })
  duration?: string;

  @Metadata({ data: "json, name=price" })
  price?: number;

  @Metadata({ data: "json, name=price_unit" })
  priceUnit?: string;

  @Metadata({ data: "json, name=recording_sid" })
  recordingSid?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=status" })
  status?: RecordingTranscriptionEnumStatusEnum;

  @Metadata({ data: "json, name=transcription_text" })
  transcriptionText?: string;

  @Metadata({ data: "json, name=type" })
  type?: string;

  @Metadata({ data: "json, name=uri" })
  uri?: string;
}